Authentication by user's group

+++ src/main/java/org/archivemanager/portal/web/DigitalReserveController.java
@@ -11,6 +11,8 @@ import org.heed.openapps.User;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.heed.openapps.search.data.RestData;
 import org.heed.openapps.entity.data.FormatInstructions;
+import org.heed.openapps.security.SecurityService;
+import org.heed.openapps.Group;
 
 import org.heed.openapps.data.RestResponse;
 import org.heed.openapps.entity.EntityService;
@@ -23,6 +25,7 @@ import org.springframework.web.bind.annotation.ResponseBody;
 
 import com.liferay.portal.kernel.bean.BeanLocator;
 import com.liferay.portal.kernel.bean.PortletBeanLocatorUtil;
+import com.liferay.portal.util.PortalUtil;
 
 
 @Controller
@@ -30,10 +33,10 @@ import com.liferay.portal.kernel.bean.PortletBeanLocatorUtil;
 public class DigitalReserveController {
     private EntityService entityService;
     private SearchService searchService;
+    private SecurityService securityService;
 
-    //@Autowired
-    //User user;
-
+    // @Autowired
+    // User user;
 
     User user = new User();
     FormatInstructions inst = new FormatInstructions();
@@ -43,7 +46,7 @@ public class DigitalReserveController {
     //a get request that returns all the current subjects
     @ResponseBody
     @RequestMapping(value="/subject/view", method = RequestMethod.GET)
-    public RestResponse<Subject> viewSubjects() throws Exception{
+    public RestResponse<Subject> viewSubjects(HttpServletRequest request, HttpServletResponse response) throws Exception{
       RestResponse<Subject> data = new RestResponse();
         entityService = getEntityService();
         Entity ent = entityService.getEntity(11);
@@ -51,20 +54,32 @@ public class DigitalReserveController {
         return data;
     }
 
+    //just for testing: checking user's data
+    @ResponseBody
+    @RequestMapping(value="/user", method = RequestMethod.GET)
+    public String viewUser(HttpServletRequest request, HttpServletResponse response) throws Exception{
+      String result = String.valueOf(checkEducator(request));
+      return result;
+    }
+
     //returns the new made subject or just returns the corresponding subject
     //going to be changed to post
     @ResponseBody
     @RequestMapping(value="/subject/add", method = RequestMethod.GET)
-    public void addSubject(@RequestParam(required=true) String sub){
+    public void addSubject(HttpServletRequest request, HttpServletResponse response, @RequestParam(required=true) String sub)throws Exception{
+      if(checkEducator(request)){
         dummy.addSubject(sub);
+      }
     }
 
     //deletes a whole subject
     //going to be changed to post
     @ResponseBody
     @RequestMapping(value="/subject/delete", method = RequestMethod.GET)
-    public void deleteSubject(@RequestParam(required=true) String sub){
+    public void deleteSubject(HttpServletRequest request, HttpServletResponse response, @RequestParam(required=true) String sub)throws Exception{
+      if(checkEducator(request)){
         dummy.deleteSubject(sub);
+      }
     }
 
     //add entity to a subject
@@ -75,13 +90,16 @@ public class DigitalReserveController {
                           @RequestParam(required=true) Long id,
                           @RequestParam(required=true) String sub) throws Exception {
         //add the entity to the list
-        entityService = getEntityService();
-        Entity ent = entityService.getEntity(id);
-        Object obj = entityService.export(inst, ent);
-        dummy.addItem(sub, obj);
         RestResponse data = new RestResponse();
-        data.getResponse().addData(obj);
+        if(checkEducator(request)){
+          entityService = getEntityService();
+          Entity ent = entityService.getEntity(id);
+          Object obj = entityService.export(inst, ent);
+          dummy.addItem(sub, obj);
+          data.getResponse().addData(obj);
+        }
         return data;
+
     }
 
     //delete the entity that belongs in a subject
@@ -92,10 +110,12 @@ public class DigitalReserveController {
                                @RequestParam(required=true) Long id,
                                @RequestParam(required=true) String sub) throws Exception {
         //delete the entity to the list
-        entityService = getEntityService();
-        Entity ent = entityService.getEntity(id);
-				//how the subject looks like now
-				Subject updated = dummy.deleteItem(sub, entityService.export(inst, ent));
+        if(checkEducator(request)){
+          entityService = getEntityService();
+          Entity ent = entityService.getEntity(id);
+  				//how the subject looks like now
+  				Subject updated = dummy.deleteItem(sub, entityService.export(inst, ent));
+        }
     }
 
 	public SearchService getSearchService() {
@@ -116,4 +136,30 @@ public class DigitalReserveController {
 		}
 		return entityService;
 	}
+  public SecurityService getSecurityService() {
+    if(securityService == null) {
+      BeanLocator locator = PortletBeanLocatorUtil.getBeanLocator("archivemanager-portlet");
+      securityService = (SecurityService)locator.locate("securityService");
+    }
+    return securityService;
+  }
+
+  private boolean checkEducator(HttpServletRequest request) throws Exception{
+    HttpServletRequest httpReq2 = PortalUtil.getOriginalServletRequest(request);
+    User user2 = getSecurityService().getCurrentUser(request);
+    List<Group> groups = user2.getGroups();
+    int len = groups.size();
+    if(len != 0){
+      for(int i =0; i < len; i++){
+        Group grp = groups.get(i);
+        String nm = grp.getName();
+        if(nm.equals("Educators")){
+          return true;
+        }
+      }
+    }
+    return false;
+  }
+
+
 }
